pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "grammar-enforcement"
version = "1.0.0"
description = "Grammar-constrained refinement of safety operational rules for autonomous systems (SEAMS 2026)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SEAMS 2026 Team"},
]
keywords = [
    "operational-rules",
    "grammar-enforcement",
    "autonomous-systems",
    "CBF",
    "safety",
    "SEAMS",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "lark>=1.0,<2.0",
    "numpy>=1.20,<3.0",
    "scikit-learn>=1.0,<2.0",
    "pandas>=1.3,<3.0",
    "openpyxl>=3.0,<4.0",
    "pyyaml>=6.0,<7.0",
    "colorama>=0.4,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<8.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
ui = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "httpx>=0.24",
]

[project.scripts]
run-pipeline = "run_pipeline:main"

[project.urls]
Repository = "https://github.com/seams2026/grammar-enforcement"

# ─── Tool configurations ─────────────────────────────────────────────

[tool.setuptools.packages.find]
include = [
    "core*",
    "parsers*",
    "validators*",
    "semantic*",
    "minimality*",
    "data*",
    "shared*",
    "rule_inference*",
    "rule_validation*",
    "cbf_data*",
    "examples*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.coverage.run]
source = [
    "core",
    "parsers",
    "validators",
    "semantic",
    "minimality",
    "data",
    "shared",
    "rule_inference",
    "rule_validation",
    "cbf_data",
]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
fail_under = 50
